requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
comm==0.2.3
cryptography==46.0.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-genai==1.73.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
narwhals==2.19.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
Send2Trash==2.1.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.56.0
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
